More code cleanups
- Moved cache functions to their own package
- Some other minor changes
