+++ playcert/__init__.py
@@ -2,7 +2,8 @@ from pyramid.config import Configurator
 
 
 def main(global_config, **settings):
-    """ This function returns a Pyramid WSGI application.
+    """
+    This function returns a Pyramid WSGI application.
     """
     config = Configurator(settings=settings)
     config.include('pyramid_chameleon')
